Enable the hybrid DTID
We would like to eliminate the guest periodic timer events and
allows the occasional aperiodic timer events.

To eliminate the guest periodic timer events, it is handled by
the guest OS modification and helps from the hypervisor. The
guest OS is modified to ignore the periodic timer events by
omitting the TMICT MSR-writes. The hypervisor sets the PIR
timer-interrupt bit and ON bit, while the guest LAPIC fires
the PIN at the guest periodic timer interval.

For the guest aperiodic timer events, we allow them to issue
TMICT MSR-writes. After the control is transferred to the
hypervisor, the KVM sets the PIR timer-interrupt bit, ON bit
and configure the TMICT by the guest aperiodic timer interval.

modified:   arch/x86/include/asm/osnet.h