Investigate the NIC-induced timer interrupts
Using the existing posted-interrupt approach to directly
deliver the periodic timer interrupt to the guest has
encountered a problem of NIC-induced timer interrupt.

To deliver the timer interrupt to the guest, the PIR
timer-interrupt and ON bit are set before the LAPIC timer of
guest core fires the posted-interrupt notification. Our
exiting algorithm sets the proper PIR and ON bit ahead of
time, but may not set them just right before PIN arrives. Two
hardware chips generate the PIN: LAPIC timer chip (our
approach) and VT-d (meant for the I/O devices). If the PIN
generated by the passthrough I/O devices arrives before the
PIN generated by the LAPIC timer chip of guest core, the
NIC-included time interrupt happens. This is the spurious
timer interrupt for the guest. It results in the early and
unnecessary direct delivery of timer interrupt to the guest.
The guest needs to schedule the next timer interrupt.

For example, the PIR and ON bit are set just right after the
previous guest timer interrupt. This leaves the large
vulnerable window for the NIC-induced PINs delivered to the
guest by the VT-d chipset. When the guest runs the
network-intensive application with the high-performance NIC,
the number of unwanted timer interrupt increases dramatically.
When it is time to handle the NIC interrupt, the guest handles
the spurious timer interrupt. The guest is not able to handle
the NIC interrupts on time.  Thus, the high number of
NIC-induced timer interrupts reduce the network-bandwidth
performance.

modified:   arch/x86/include/asm/osnet.h
modified:   arch/x86/kernel/apic/apic.c
modified:   arch/x86/kernel/irq.c
new file:   osnet/irq.c