Emulate the guest periodic timer using the host hrtimer
For each vCPU, the periodic timer events are emulated by the
corresponding hrtimer pinned to the hypervisor core. Before
starting the timer, the hypervisor core performs the initial
calibration. It sets the PIR timer-interrupt and ON bit for
the target vCPU, polls until the proper bits are cleared and
starts the timer. When it is about 50 us before the LAPIC
timer of guest core fires the PIN, the timer is up. Its
callback function does two things. First, it sets the PIR
timer-interrupt and ON bit. Second, it starts the next timer
based on the previous expiration and guest periodicity.

modified:   arch/x86/include/asm/osnet.h
modified:   arch/x86/kernel/apic/apic.c
modified:   virt/kvm/kvm_main.c